+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @computesdk/modal
 
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies [b027cd9]
+  - computesdk@1.9.5
+
+## 1.7.0
+
+### Minor Changes
+
+- 64569f1: implement ports update for modal
+
+### Patch Changes
+
+- computesdk@1.9.4
+
+## 1.6.2
+
+### Patch Changes
+
+- Updated dependencies [f38470d]
+- Updated dependencies [f38470d]
+  - computesdk@1.9.3
+
+## 1.6.1
+
+### Patch Changes
+
+- Updated dependencies [1ac5ad2]
+  - computesdk@1.9.1
+
 ## 1.6.0
 
 ### Minor Changes