+++ packages/render/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @computesdk/render
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [b027cd9]
+  - computesdk@1.9.5
+
+## 1.1.3
+
+### Patch Changes
+
+- computesdk@1.9.4
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [f38470d]
+- Updated dependencies [f38470d]
+  - computesdk@1.9.3
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [1ac5ad2]
+  - computesdk@1.9.1
+
 ## 1.1.0
 
 ### Minor Changes